chore: switch to generated corev1 ctrls
